pretrain_comparison/requirements.txt
click==8.1.7
einops==0.8.1
h5py==3.11.0
ipykernel==6.26.0
ipywidgets==8.1.1
loguru==0.7.2
matplotlib==3.8.0
notebook==7.0.6
numpy==1.26.4
pandas==2.2.3
pyedflib==0.1.37
PyYAML==6.0.2
python-dotenv==1.0.0
scikit_learn==1.3.1
scipy==1.15.3
seaborn==0.13.2
statsforecast==1.6.0
statsmodels==0.14.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.66.2
wandb==0.16.
ruamel.yaml>=0.18.10
lifelines>=0.30.0
